Extract useToken as separate custom hook;
